Swap Elena for Rowan
Rowan is a similar medium x-height serif, targeted towards long passages of text. I don't like the uppercase forms as much (the serifs are more pronounced than those of Elena's), but Rowan is generously licensed and is variable weight. There were many places where I had needed a medium weight of Elena, but had to settle for bold.
I've adjusted the sizes to keep the baseline the same for now